Change name for message-item component
